refactor(tasks): give the skills picker its own file
The three pickers pushed CronJobConfigurationPickers.swift past the
file-size warning. The skills picker is the natural split: it is the only
one that commits a collection. Pure move, no behavior change.
